pyproject.toml
[project]
name = "ai-auto-trader"
version = "0.1.0"
description = ""
authors = [
    {name = "nno0obb",email = "nno0obb@kakao.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai (>=1.97.1,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pyupbit (>=0.2.34,<0.3.0)",
    "requests (>=2.32.4,<3.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "ta (>=0.11.0,<0.12.0)",
    "redis (>=6.4.0,<7.0.0)",
    "gitpython (>=3.1.45,<4.0.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
